backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
python-dateutil==2.8.2 